requirements.txt
fastapi
uvicorn
pytesseract
pillow
python-multipart
torch
numpy
